Refactored tsquery parser to handle quoted strings, exclude NOT terms
ts_rank() ignores operators which leads to contradictory results when
NOT operators are used. As a result the TSQuery parser will output two
strings: one same as before and the other with all NOT terms removed.

The infix notation used by the parser did not work with the quoted
strings so it has been changed to use recursive terms instead. The
parser and web forms now validate strings by using sets of extended
ASCII characters and excluding other characters.

Logging added to search results.