Refactored modify_stop_areas()
The functions have been refactored to to deal with queries hanging;
selecting the mode of the stop localities are now handled on the Python
side. For any stop areas with ambiguous localities, the closest
localities will be selected.

All tram stops and areas take admin area ref from the localities
they are contained within.