Moved populate subpackage functionality to a separate module
Like with models subpackage, the public functions for committing new
data is now imported into the subpackage namespace.

When the backup option is enabled for populating data, the database is
no longer restored on reaching an error.

The default config option for ATCO codes covering all areas is now
None.