Added panel for live bus stop times
Panel will show a message reminding to zoom in first before seeing
stops and will show stop and live times when a marker is clicked on,
instead of linking to the stop page in question.

Base HTML template has been modified a bit to allow for removal and
moving of footer into the panel.