Moved map JS to external file and use tile-based streaming
Each map tile load is followed by another tile for stop point markers,
and vice versa for unloading. This should be more responsive than using
bounds to calculate which stops to show.

The minimum screen width for the map panel has been increased to 800px
for maps as the remaining space for the map is too small.

Global constants are now used by Flask to pass values such as URLs.