Merge all services sharing line names and sequences of stops
Using temporary tables, services are compared on stops they cover and
if they share more than half of stops and use the same line name they
are very likely to be the same service (which may be run by different
operators) and can be merged by updating journey pattern references and
deleting the extra services. Since 'outbound' and 'inbound' directions
can be different for different services especially if they cover
multiple regions, the stops served in different directions are compared
and their directions swapped as necessary.