Cyclic path fixes for graph layout
Another graph added to tests with lots of cycles. Line ordering is made
optional for drawing but mostly used for testing anyway.

Fixes:

- When adding starting cycles to the layout, the lower of the starting
column or the column to the right of the ending lines is picked.

- The vertex, incoming lines, and starting cycles on the same row as
ending cycles being added are moved if they are placed to the right of
the next row's vertex where the ending cycles would be added.

- Columns for starting cycles are stored on the preceding row - that row
needs to be accessed when rearranging columns.