Integrated services into map and refactored map views
A ServicesData object stores service data and issues callbacks when
services are retrieved, and loaded into a RouteLayer object whose layer
is added to map. The map has a update() method which updates loads or
clears services and stops as needed and updates the stop panel.

Map is initialised using a object containing stop and service data.

Map views use a single function with multiple routes for services, stops
and coordinates to make it easier to specify which URL should be used
(eg service and stop but no direction). Client side URL updating has
been modified to ensure pages are always reproducible on reload.

Added fixes in the diagram drawing code.