Made timetables work with timezone changes
Departure times are checked to see if they fall within 0100 and 0200 on
the last Sunday of March and October and are skipped over or repeated in
different timezones.

Subquerys are used instead of CTEs and bind parameters are used in the
timetable query to reduce repetition of values such as the user-selected
date.

A clause for VALUES has been added so it can be used in SQLAlchemy
queries.