Cleaned up relationships & queries and set relationships to raise
The locality reference for StopPoint is sometimes used before any
queries which can raise an error as it was implicitly declared by the
relationship's backref attribute. The relationship is now declared from
both sides so this shouldn't happen.

All relationships have the lazy attribute to 'raise', ie, an error will
be raised if a relationship attribute was called but was not loaded as
part of the initial query.