Added NOC dataset
The National Operator Code dataset contains data on all public transport
operators, including contact info. This is more comphrensive than what
is provided from TNDS but there are still a few operators not covered by
the NOC dataset, so operator data will be extracted from TNDS although
will not be added if data already exists.

With most of new data added to the Operator model, it will be used over
LocalOperator which is more of an association table.

The service modes have been expanded to include coach and metro /
underground.

The populate command had NOC options added and arguments changed to
keywords which should be more readable.