Removed item magic methods; graph sequence and paths are cached
The __getitem__, __setitem__ and __delitem__ methods are removed from
Graph as they're not explicit about what they do - it's easier to use
add_vertex, remove_vertex, etc.

The adjacency list is checked and if it is the same the paths and
sequence from the last time the graph is analysed are used.

_Layout gets a copy_from() method which simplifies copying of data while
drawing graph.