Changed Graph.search_paths() to look for preceding vertices directly
Previously, Graph.search_paths() will start with other vertices and look
for the shortest path to a particular vertex. Now, paths will be
transversed backwards using Graph.preceding(). This doesn't give the
same results and I'm not sure if it looks as good but it is
significantly faster especially on larger graphs.