Used generate_series to repeat or skip departures in timetables
Dates and departure times were evaluated on whether they occurred within
the timezone change as part of DST and were skipped or repeated.
PostgreSQL can do this automatically in a generate_series set between
two timestamps, so by generating a hourly series with the starting and
ending timestamps to be 1 hour before and after the departure time, then
excluding these start and end times, should create 0 or 2 rows when the
timezone changes and 1 row otherwise.