Fixed issues with particular search queries
A Empty operator class has been added to cover any queries with empty
parentheses, eg 'road ()'. It always returns an empty string and is not
defined. FollowedBy will not be defined if the quoted string is empty.

When the parse tree is converted to a string with to_string() all empty
strings such as above will be skipped. The 'raise_undefined' parameter
is added to this method to make raising SearchNotDefined optional.

Operators such as 'and' at the start of search queries are excluded.