Replaced tsquery parser with PostgreSQL's own websearch_to_tsquery()
This function, new to PostgreSQL 11, replicates boolean and quoting
functionality found in other search engines. The querytree() function is
used to strip unbounded parts so it is suitable for ts_rank() which
still ignores NOT terms.

Changed search_code() to only query stop point if search query matches
an alphanumeric word of a certain length.