Added map shortcut and rearranged starred stop dependencies
StarredStops, StarredStopList uses dependency injection to
set up object references which should help list of starred stops link
back to the map on the same page, as well as resetting focusable
elements on overlay when a new list is created.

Map icon added as inline SVG to starred stop lists and search results
with direct link to map.