Better callback and error handling for StarredStops
Each of the methods have a onSuccess and a onFail callback, the former
called with data when request is successful and the latter if it fails,
either because of network issues or the response had the wrong HTTP code
(eg 404). The SMS code needs to be sent with each method call, although
the overlay still needs it to be stored so the add() method can be
called twice.