Replaced service merging with pairs of similar services
Merging services by comparing their stops worked in most cases but there
were a few where the patterns were completely different, eg one with a
circular pattern and another with a pair of patterns, both covering the
exact same route. This would have made the timetable and diagram
needlessly complicated.

Instead, services were compared on their stops and added to a new
service_pair table and with a threshold the most similar services can
be shown for each service.

It would still be good to do some merging as there are several services
that are basically identical.