Fixed timetable data displayed in wrong time zone when server uses UTC
Journey departure times are always local time, but departure times
generated to check for missing/duplicate departures when switching
between GMT and BST are not. Not a problem when local machine is
uses the same time zone but on server with UTC the wrong times will be
picked every time. Fixed by casting non-local departure to local time
before comparsion.

Set time zones before running queries for the DST tests to catch this.