Use a generated code for services
A code related to line and description is much more readable than a
numeric ID and should be more stable through repopulation as long as
file order does not change.

A shorter description generated from the given description is used for
the code and saved in a new column. If no description exists the origin
and destination from the associated standard service is used. This
description is also used for lists of services.

For example, a service 120 with description 'Halfway - Fulwood' will be
given code '120-halfway-fulwood', the second service
'120-halfway-fulwood-2' and so on.

Added migration which generates these codes and extra XSLT functionality
to do the same during population.

Fixed handling of other types for arguments passed in xslt_text_func.

Renamed 'short_ind' in FTS column to 'indicator'.

fixup: short description set to not remove words