Moved static files back to /static but added root static endpoint
The problem with setting static_url_path to '/' is it breaks 405
responses, as the static endpoint will capture all GET requests before
Flask can check if other HTTP methods exist for these paths.

Moving the static endpoint back to '/static' and duplicating the
endpoint with use of an URL converter to check whether files exist is a
little messy but it works.