Restyled website to mix of brutalist and original aesthetic
The original style had a grey background with white boxes for sections.
The brutalist style didn't work out too well because of the information
layout and accessibility shouldn't be compromised. Most of the elements
kept their appearance and the colours were unchanged but the header is
now black, the background a light grey and all sections white.

Restyled page numbers for search results.