Changed live data interval to be relative to current timestamp
JS functionality on mobile phones can be suspended when the user leaves
the browser app. On returning to a page with live data the interval will
resume with the same number of seconds which isn't ideal. Switching to
an expiration timestamp based on Date.now() plus interval should help as
the live data will update automatically when the user returns to the
page after some time.