Fixed issues with docker + poetry
Poetry does a few things differently and support for system installs
is not as robust as pipenv's. The application will run locally with
poetry installing dependencies in a virtualenv. The nxb command has
been removed in favour of a manage.py script which can then run within
the virtualenv created by poetry.