Refactored population to use a single transaction
The main problem with running multiple transactions is that in the
event of an error the whole database has to be restored. Moving to run
everything within a single transaction solves this, but required
refactoring every component to add the connection as a dependency and
move as much as possible into SQLAlchemy's expression API. The
population logic was moved to a runner function outside of the CLI.

Indexes also had to be dropped at start and recreated at end to save on
storage but had the benefit of speeding up the population process.