Improve timetable performance with generated data
The current query for timetables has a lot of problems - several tables
need to be joined for scheduling journeys which is difficult for the
query planner to optmise for and the times need to be generated from
summing timings cumulatively after sorting.

Create a JSON 'data' column on the journey table to hold an array of
stops and their timings, eliminating the need for a subquery. Rewrite
the journey scheduling subquery to aggregate bank holidays/operating
periods before joining, which reduces the number of potential rows
considerably.